Don't nudgePeriodically anymore
This eliminates the annoying performance consideration, meaning the high-level
Clock API can be used in more situations (e.g. implementing updatable timeout).

Besides, this technique does not guard against another theoretical possibility:
the computer or the process is asleep for more than 49.7 days.

I plan to add support for longer wraparound, where available.  For example, on
Windows, we will use QueryPerformanceCounter if it is available, and fall back
to GetTickCount if it is not.