Add support for GetTickCount64, and default to GetTickCount64 or GetTickCount instead of QueryPerformanceCounter
QueryPerformanceCounter is less stable in the long run, and stops counting (or
maybe counts slower) when the computer is suspended.  'GetTickCount' does not
have this problem.

However, clock_gettime(CLOCK_MONOTONIC) also stops running when the computer is
suspended.  I may just list this as a "known issue" in the module documentation.