Extend SystemClock to support an internal cumulative time that does not truncate precision
Still need to update the Windows drivers and test them.

This fixes a theoretical accuracy loss issue when using a time source where
systemClockDiffTime looks something like this:

    (new - old) / frequency

This is the case with QueryPerformanceCounter (not used by default), and it is
the case on Mac OS X when using mach_absolute_time and mach_timebase_info
(not supported yet).